Ensure that the application correctly counts the number of sandwich collections
Cast the count result from the database to an integer type in `getSandwichCollectionsCount` to avoid type mismatch errors.

+++ server/database-storage.ts
@@ -169,7 +169,7 @@ export class DatabaseStorage implements IStorage {
   }
 
   async getSandwichCollectionsCount(): Promise<number> {
-    const result = await db.select({ count: sql<number>`count(*)` }).from(sandwichCollections);
+    const result = await db.select({ count: sql<number>`count(*)::int` }).from(sandwichCollections);
     return Number(result[0].count);
   }
